Create Exercise 4 animated drawing
a black circle is drawn when you move the mouse (all previous black circles are cleared).

the x coordinate for the circle is the mouseY coordinate and the y coordinate for the circle is mouseX.

the size of the circle is determined by the x mouse coordinate.

When you run your code, the black circle should overlap the white circle when you drag your mouse from (0,0) diagnolally down to (100,100). When you move your mouse horizontally, the black circle show grow/shrink in size. It remains the same size when the mouse is dragged vertically. Note that the white circle's behaviour hasn't changed.